+++ to_do_app/lib/app_router.dart
@@ -1,9 +1,10 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
-import 'addTask/newTask.dart';
-import 'detailTask/detailTask.dart';
-import 'home/myHomePage.dart';
+import 'addTask/new_task.dart';
+import 'detailTask/detail_task.dart';
+import 'domain/card_to_do.dart';
+import 'home/my_home_page.dart';
 
 part 'app_router.gr.dart';
