+++ to_do_app/lib/home/list_Item.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:to_do_app/domain/card_to_do.dart';
+
+class ListItem extends StatelessWidget {
+  final _biggerFont = const TextStyle(fontSize: 18);
+  CardToDo element;
+  final Function(bool?) onChanged;
+  final Function() viewDetail;
+  ListItem({Key? key, required this.element, required this.onChanged, required this.viewDetail}) : super(key: key);
+
+
+  @override
+  Widget build(BuildContext context) => GestureDetector(
+      onTap: viewDetail,
+      child: ListTile(
+        title: Text(
+          element.title,
+          style: _biggerFont,
+        ),
+        subtitle: Text(element.description),
+        trailing: Checkbox(
+          checkColor: Colors.white,
+          activeColor: Colors.pinkAccent,
+          value: element.done,
+          onChanged: onChanged
+        ),
+      ),
+    );
+  }
+
